+++ batched/dense/src/KokkosBatched_Gemm_Decl.hpp
@@ -51,73 +51,6 @@
 #include <KokkosKernels_Error.hpp>
 
 namespace KokkosBatched {
-/********************* BEGIN functor-level routines *********************/
-///
-/// Serial Gemm
-///
-
-template <typename ArgTransA, typename ArgTransB, typename ArgAlgo>
-struct SerialGemm {
-  template <typename ScalarType, typename AViewType, typename BViewType,
-            typename CViewType>
-  KOKKOS_INLINE_FUNCTION static int invoke(const ScalarType alpha,
-                                           const AViewType &A,
-                                           const BViewType &B,
-                                           const ScalarType beta,
-                                           const CViewType &C);
-};
-
-///
-/// Team Gemm
-///
-
-template <typename MemberType, typename ArgTransA, typename ArgTransB,
-          typename ArgAlgo>
-struct TeamGemm {
-  template <typename ScalarType, typename AViewType, typename BViewType,
-            typename CViewType>
-  KOKKOS_INLINE_FUNCTION static int invoke(
-      const MemberType &member, const ScalarType alpha, const AViewType &A,
-      const BViewType &B, const ScalarType beta, const CViewType &C);
-};
-
-///
-/// TeamVector Gemm
-///
-
-template <typename MemberType, typename ArgTransA, typename ArgTransB,
-          typename ArgAlgo>
-struct TeamVectorGemm {
-  template <typename ScalarType, typename AViewType, typename BViewType,
-            typename CViewType>
-  KOKKOS_INLINE_FUNCTION static int invoke(
-      const MemberType &member, const ScalarType alpha, const AViewType &A,
-      const BViewType &B, const ScalarType beta, const CViewType &C);
-};
-
-///
-/// Selective Interface
-///
-template <typename MemberType, typename ArgTransA, typename ArgTransB,
-          typename ArgMode, typename ArgAlgo>
-struct Gemm {
-  template <typename ScalarType, typename AViewType, typename BViewType,
-            typename CViewType>
-  KOKKOS_FORCEINLINE_FUNCTION static int invoke(
-      const MemberType &member, const ScalarType alpha, const AViewType &A,
-      const BViewType &B, const ScalarType beta, const CViewType &C) {
-    int r_val = 0;
-    if (std::is_same<ArgMode, Mode::Serial>::value) {
-      r_val = SerialGemm<ArgTransA, ArgTransB, ArgAlgo>::invoke(alpha, A, B,
-                                                                beta, C);
-    } else if (std::is_same<ArgMode, Mode::Team>::value) {
-      r_val = TeamGemm<MemberType, ArgTransA, ArgTransB, ArgAlgo>::invoke(
-          member, alpha, A, B, beta, C);
-    }
-    return r_val;
-  }
-};
-/********************* END functor-level routines *********************/
 
 /********************* BEGIN non-functor-level routines *********************/
 
@@ -647,8 +580,6 @@ int BatchedGemm(BatchedGemmHandleType *const handle, const ScalarType alpha,
 }  // namespace KokkosBatched
 
 #include "KokkosBatched_Gemm_Serial_Impl.hpp"
-#include "KokkosBatched_Gemm_Team_Impl.hpp"
-#include "KokkosBatched_Gemm_TeamVector_Impl.hpp"
 #include "KokkosBatched_Gemm_DblBuf_Impl.hpp"
 #include "KokkosBatched_Gemm_Armpl_Impl.hpp"
