+++ batched/dense/unit_test/Test_Batched_SerialInverseLU.hpp
@@ -6,8 +6,7 @@
 
 //#include "KokkosBatched_Vector.hpp"
 
-#include "KokkosBatched_Gemm_Decl.hpp"
-#include "KokkosBatched_Gemm_Serial_Impl.hpp"
+#include "KokkosBlas3_gemm.hpp"
 #include "KokkosBatched_LU_Decl.hpp"
 #include "KokkosBatched_LU_Serial_Impl.hpp"
 #include "KokkosBatched_InverseLU_Decl.hpp"
@@ -47,8 +46,9 @@ struct Functor_BatchedSerialGemm {
 
     for (int i = 0; i < static_cast<int>(aa.extent(0)); ++i) aa(i, i) += 10.0;
 
-    SerialGemm<typename ParamTagType::transA, typename ParamTagType::transB,
-               AlgoTagType>::invoke(_alpha, aa, bb, _beta, cc);
+    KokkosBlas::SerialGemm<typename ParamTagType::transA,
+                           typename ParamTagType::transB,
+                           AlgoTagType>::invoke(_alpha, aa, bb, _beta, cc);
   }
 
   inline void run() {
