+++ batched/dense/unit_test/Test_Batched_TeamGemm.hpp
@@ -6,10 +6,7 @@
 
 //#include "KokkosBatched_Vector.hpp"
 
-#include "KokkosBatched_Gemm_Decl.hpp"
-#include "KokkosBatched_Gemm_Serial_Impl.hpp"
-#include "KokkosBatched_Gemm_Team_Impl.hpp"
-
+#include "KokkosBlas3_gemm.hpp"
 #include "KokkosKernels_TestUtils.hpp"
 
 using namespace KokkosBatched;
@@ -45,10 +42,10 @@ struct Functor_TestBatchedTeamGemm {
     auto bb = Kokkos::subview(_b, k, Kokkos::ALL(), Kokkos::ALL());
     auto cc = Kokkos::subview(_c, k, Kokkos::ALL(), Kokkos::ALL());
 
-    KokkosBatched::TeamGemm<MemberType, typename ParamTagType::transA,
-                            typename ParamTagType::transB,
-                            AlgoTagType>::invoke(member, _alpha, aa, bb, _beta,
-                                                 cc);
+    KokkosBlas::TeamGemm<MemberType, typename ParamTagType::transA,
+                         typename ParamTagType::transB,
+                         AlgoTagType>::invoke(member, _alpha, aa, bb, _beta,
+                                              cc);
   }
 
   inline void run() {
