+++ blas/impl/KokkosBlas3_serial_gemm_internal.hpp
@@ -1,16 +1,59 @@
-#ifndef __KOKKOSBATCHED_GEMM_SERIAL_INTERNAL_HPP__
-#define __KOKKOSBATCHED_GEMM_SERIAL_INTERNAL_HPP__
+/*
+//@HEADER
+// ************************************************************************
+//
+//                        Kokkos v. 3.0
+//       Copyright (2020) National Technology & Engineering
+//               Solutions of Sandia, LLC (NTESS).
+//
+// Under the terms of Contract DE-NA0003525 with NTESS,
+// the U.S. Government retains certain rights in this software.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are
+// met:
+//
+// 1. Redistributions of source code must retain the above copyright
+// notice, this list of conditions and the following disclaimer.
+//
+// 2. Redistributions in binary form must reproduce the above copyright
+// notice, this list of conditions and the following disclaimer in the
+// documentation and/or other materials provided with the distribution.
+//
+// 3. Neither the name of the Corporation nor the names of the
+// contributors may be used to endorse or promote products derived from
+// this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY NTESS "AS IS" AND ANY
+// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+// PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL NTESS OR THE
+// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+// Questions? Contact Siva Rajamanickam (srajama@sandia.gov)
+//
+// ************************************************************************
+//@HEADER
+*/
+
+#ifndef KOKKOSBLAS3_GEMM_SERIAL_INTERNAL_HPP
+#define KOKKOSBLAS3_GEMM_SERIAL_INTERNAL_HPP
 
 /// \author Kyungjoo Kim (kyukim@sandia.gov)
 
-#include "KokkosBatched_Util.hpp"
-
+#include "KokkosBlas_util.hpp"
 #include "KokkosBlas1_set_impl.hpp"
 #include "KokkosBlas1_serial_scal_impl.hpp"
+#include "KokkosBlas3_gemm_inner_fix.hpp"
 
-#include "KokkosBatched_InnerGemmFixC_Serial_Impl.hpp"
-
-namespace KokkosBatched {
+namespace KokkosBlas {
+namespace Impl {
 
 ///
 /// Serial Internal Impl
@@ -89,7 +132,8 @@ KOKKOS_INLINE_FUNCTION int SerialGemmInternal<Algo::Gemm::Blocked>::invoke(
     if (m <= 0 || n <= 0 || k <= 0) return 0;
     const ValueType alpha_value(alpha);
 
-    InnerGemmFixC<mbAlgo, nbAlgo> inner(as0, as1, bs0, bs1, cs0, cs1);
+    KokkosBlas::InnerGemmFixC<mbAlgo, nbAlgo> inner(as0, as1, bs0, bs1, cs0,
+                                                    cs1);
     auto gemm = [&](const int ib, const int jb, const int pb,
                     const ValueType *KOKKOS_RESTRICT AA,
                     const ValueType *KOKKOS_RESTRICT BB,
@@ -132,6 +176,7 @@ KOKKOS_INLINE_FUNCTION int SerialGemmInternal<Algo::Gemm::Blocked>::invoke(
   return 0;
 }
 
-}  // namespace KokkosBatched
+}  // namespace Impl
+}  // namespace KokkosBlas
 
 #endif
