+++ perf_test/blas/blas3/KokkosBlas3_gemm_perf_test.hpp
@@ -54,9 +54,6 @@
 #include <KokkosBlas3_gemm.hpp>
 
 #include "KokkosBatched_Gemm_Decl.hpp"
-#include "KokkosBatched_Gemm_Serial_Impl.hpp"
-//#include "KokkosBatched_Gemm_Team_Impl.hpp"
-//#include "KokkosBatched_Gemm_TeamVector_Impl.hpp"
 #include "KokkosBatched_Util.hpp"
 #include "gtest/gtest.h"  // EXPECT_NEAR
 #include "KokkosKernels_TestUtils.hpp"
@@ -418,7 +415,7 @@ void __do_gemm_serial_batched_template(options_t options,
           C = Kokkos::subview(_gemm_args.C, Kokkos::ALL(), Kokkos::ALL(), j);
         }
 
-        KokkosBatched::SerialGemm<TransAType, TransBType, AlgoType>::invoke(
+        KokkosBlas::SerialGemm<TransAType, TransBType, AlgoType>::invoke(
             _gemm_args.alpha, A, B, _gemm_args.beta, C);
       }
     }
@@ -615,7 +612,7 @@ struct parallel_batched_gemm_range_policy {
     auto svB = Kokkos::subview(gemm_args_.B, i, Kokkos::ALL(), Kokkos::ALL());
     auto svC = Kokkos::subview(gemm_args_.C, i, Kokkos::ALL(), Kokkos::ALL());
 
-    KokkosBatched::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
+    KokkosBlas::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
         gemm_args_.alpha, svA, svB, gemm_args_.beta, svC);
   }
 
@@ -625,7 +622,7 @@ struct parallel_batched_gemm_range_policy {
     auto svB = Kokkos::subview(gemm_args_.B, Kokkos::ALL(), Kokkos::ALL(), i);
     auto svC = Kokkos::subview(gemm_args_.C, Kokkos::ALL(), Kokkos::ALL(), i);
 
-    KokkosBatched::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
+    KokkosBlas::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
         gemm_args_.alpha, svA, svB, gemm_args_.beta, svC);
   }
 
@@ -638,7 +635,7 @@ struct parallel_batched_gemm_range_policy {
     auto svC =
         Kokkos::subview(gemm_args_.Cv.vec_3d, i, Kokkos::ALL(), Kokkos::ALL());
 
-    KokkosBatched::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
+    KokkosBlas::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
         gemm_args_.alpha, svA, svB, gemm_args_.beta, svC);
   }
 
@@ -651,7 +648,7 @@ struct parallel_batched_gemm_range_policy {
     auto svC =
         Kokkos::subview(gemm_args_.Cv.vec_3d, Kokkos::ALL(), Kokkos::ALL(), i);
 
-    KokkosBatched::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
+    KokkosBlas::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
         gemm_args_.alpha, svA, svB, gemm_args_.beta, svC);
   }
 
@@ -700,7 +697,7 @@ struct parallel_batched_gemm {
     auto svB = Kokkos::subview(gemm_args_.B, i, Kokkos::ALL(), Kokkos::ALL());
     auto svC = Kokkos::subview(gemm_args_.C, i, Kokkos::ALL(), Kokkos::ALL());
 
-    KokkosBatched::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
+    KokkosBlas::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
         gemm_args_.alpha, svA, svB, gemm_args_.beta, svC);
   }
 
@@ -711,7 +708,7 @@ struct parallel_batched_gemm {
     auto svB = Kokkos::subview(gemm_args_.B, Kokkos::ALL(), Kokkos::ALL(), i);
     auto svC = Kokkos::subview(gemm_args_.C, Kokkos::ALL(), Kokkos::ALL(), i);
 
-    KokkosBatched::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
+    KokkosBlas::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
         gemm_args_.alpha, svA, svB, gemm_args_.beta, svC);
   }
 
@@ -722,9 +719,9 @@ struct parallel_batched_gemm {
     auto svB = Kokkos::subview(gemm_args_.B, i, Kokkos::ALL(), Kokkos::ALL());
     auto svC = Kokkos::subview(gemm_args_.C, i, Kokkos::ALL(), Kokkos::ALL());
 
-    KokkosBatched::TeamGemm<MemberType, TransAType, TransBType,
-                            BlockingType>::invoke(member, gemm_args_.alpha, svA,
-                                                  svB, gemm_args_.beta, svC);
+    KokkosBlas::TeamGemm<MemberType, TransAType, TransBType,
+                         BlockingType>::invoke(member, gemm_args_.alpha, svA,
+                                               svB, gemm_args_.beta, svC);
   }
 
   KOKKOS_INLINE_FUNCTION
@@ -734,9 +731,9 @@ struct parallel_batched_gemm {
     auto svB = Kokkos::subview(gemm_args_.B, Kokkos::ALL(), Kokkos::ALL(), i);
     auto svC = Kokkos::subview(gemm_args_.C, Kokkos::ALL(), Kokkos::ALL(), i);
 
-    KokkosBatched::TeamGemm<MemberType, TransAType, TransBType,
-                            BlockingType>::invoke(member, gemm_args_.alpha, svA,
-                                                  svB, gemm_args_.beta, svC);
+    KokkosBlas::TeamGemm<MemberType, TransAType, TransBType,
+                         BlockingType>::invoke(member, gemm_args_.alpha, svA,
+                                               svB, gemm_args_.beta, svC);
   }
 
   KOKKOS_INLINE_FUNCTION
@@ -749,11 +746,10 @@ struct parallel_batched_gemm {
     auto svC =
         Kokkos::subview(gemm_args_.C, team_idx, Kokkos::ALL(), Kokkos::ALL());
 
-    KokkosBatched::TeamVectorGemm<MemberType, TransAType, TransBType,
-                                  BlockingType>::invoke(member,
-                                                        gemm_args_.alpha, svA,
-                                                        svB, gemm_args_.beta,
-                                                        svC);
+    KokkosBlas::TeamVectorGemm<MemberType, TransAType, TransBType,
+                               BlockingType>::invoke(member, gemm_args_.alpha,
+                                                     svA, svB, gemm_args_.beta,
+                                                     svC);
   }
 
   KOKKOS_INLINE_FUNCTION
@@ -767,11 +763,10 @@ struct parallel_batched_gemm {
     auto svC =
         Kokkos::subview(gemm_args_.C, Kokkos::ALL(), Kokkos::ALL(), team_idx);
 
-    KokkosBatched::TeamVectorGemm<MemberType, TransAType, TransBType,
-                                  BlockingType>::invoke(member,
-                                                        gemm_args_.alpha, svA,
-                                                        svB, gemm_args_.beta,
-                                                        svC);
+    KokkosBlas::TeamVectorGemm<MemberType, TransAType, TransBType,
+                               BlockingType>::invoke(member, gemm_args_.alpha,
+                                                     svA, svB, gemm_args_.beta,
+                                                     svC);
   }
 
   KOKKOS_INLINE_FUNCTION
@@ -787,10 +782,9 @@ struct parallel_batched_gemm {
           auto svC = Kokkos::subview(gemm_args_.Cv.ivec_4d, i, Kokkos::ALL(),
                                      Kokkos::ALL(), vector_lane);
 
-          KokkosBatched::Gemm<MemberType, TransAType, TransBType, AlgoMode,
-                              BlockingType>::invoke(member, gemm_args_.alpha,
-                                                    svA, svB, gemm_args_.beta,
-                                                    svC);
+          KokkosBlas::Gemm<MemberType, TransAType, TransBType, AlgoMode,
+                           BlockingType>::invoke(member, gemm_args_.alpha, svA,
+                                                 svB, gemm_args_.beta, svC);
         });
   }
 
@@ -808,10 +802,9 @@ struct parallel_batched_gemm {
           auto svC = Kokkos::subview(gemm_args_.Cv.ivec_4d, vector_lane,
                                      Kokkos::ALL(), Kokkos::ALL(), i);
 
-          KokkosBatched::Gemm<MemberType, TransAType, TransBType, AlgoMode,
-                              BlockingType>::invoke(member, gemm_args_.alpha,
-                                                    svA, svB, gemm_args_.beta,
-                                                    svC);
+          KokkosBlas::Gemm<MemberType, TransAType, TransBType, AlgoMode,
+                           BlockingType>::invoke(member, gemm_args_.alpha, svA,
+                                                 svB, gemm_args_.beta, svC);
         });
   }
 
@@ -1013,7 +1006,7 @@ struct parallel_batched_gemm_experiment1 {
     auto svC = Kokkos::subview(gemm_args_.C, i, Kokkos::ALL(), Kokkos::ALL());
 
     // Uses two serial for-loops internally
-    KokkosBatched::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
+    KokkosBlas::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
         gemm_args_.alpha, svA, svB, gemm_args_.beta, svC);
   }
 };
@@ -1073,11 +1066,10 @@ struct parallel_batched_gemm_experiment2_3_4 {
 
     // Uses TeamThreadRange over C-rows
     //        ThreadVectorRange over C-cols
-    KokkosBatched::TeamVectorGemm<MemberType, TransAType, TransBType,
-                                  BlockingType>::invoke(member,
-                                                        gemm_args_.alpha, svA,
-                                                        svB, gemm_args_.beta,
-                                                        svC);
+    KokkosBlas::TeamVectorGemm<MemberType, TransAType, TransBType,
+                               BlockingType>::invoke(member, gemm_args_.alpha,
+                                                     svA, svB, gemm_args_.beta,
+                                                     svC);
   }
 
   // Experiment 3
@@ -1104,12 +1096,10 @@ struct parallel_batched_gemm_experiment2_3_4 {
           auto svC_col = Kokkos::subview(svC, Kokkos::ALL(), lane_idx);
           // TeamGemm Calls TeamThreadRange over M*N meaning the flat M*N array
           // is split over all threads of the team
-          KokkosBatched::TeamGemm<MemberType, TransAType, TransBType,
-                                  BlockingType>::invoke(member,
-                                                        gemm_args_.alpha, svA,
-                                                        svB_col,
-                                                        gemm_args_.beta,
-                                                        svC_col);
+          KokkosBlas::TeamGemm<MemberType, TransAType, TransBType,
+                               BlockingType>::invoke(member, gemm_args_.alpha,
+                                                     svA, svB_col,
+                                                     gemm_args_.beta, svC_col);
         });
   }
 
@@ -1138,12 +1128,10 @@ struct parallel_batched_gemm_experiment2_3_4 {
           auto svC_row = Kokkos::subview(svC, lane_idx, Kokkos::ALL());
           // TeamGemm Calls TeamThreadRange over M*N meaning the flat M*N array
           // is split over all threads of the team
-          KokkosBatched::TeamGemm<MemberType, TransAType, TransBType,
-                                  BlockingType>::invoke(member,
-                                                        gemm_args_.alpha,
-                                                        svA_row, svB,
-                                                        gemm_args_.beta,
-                                                        svC_row);
+          KokkosBlas::TeamGemm<MemberType, TransAType, TransBType,
+                               BlockingType>::invoke(member, gemm_args_.alpha,
+                                                     svA_row, svB,
+                                                     gemm_args_.beta, svC_row);
         });
   }
 };
@@ -1316,7 +1304,7 @@ class parallel_batched_gemm_experiment5 {
     auto svC = Kokkos::subview(C, i, Kokkos::ALL(), Kokkos::ALL());
 
     // Uses two serial for-loops internally
-    KokkosBatched::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
+    KokkosBlas::SerialGemm<TransAType, TransBType, BlockingType>::invoke(
         gemm_args.alpha, svA, svB, gemm_args.beta, svC);
   }
 };
@@ -1424,10 +1412,10 @@ class parallel_batched_gemm_experiment6 {
     auto svC = Kokkos::subview(C, i, Kokkos::ALL(), Kokkos::ALL());
 
     // Uses two serial for-loops internally
-    KokkosBatched::TeamVectorGemm<MemberType, TransAType, TransBType,
-                                  BlockingType>::invoke(member, gemm_args.alpha,
-                                                        svA, svB,
-                                                        gemm_args.beta, svC);
+    KokkosBlas::TeamVectorGemm<MemberType, TransAType, TransBType,
+                               BlockingType>::invoke(member, gemm_args.alpha,
+                                                     svA, svB, gemm_args.beta,
+                                                     svC);
   }
 };
